Fixed further issues with orientation switching in gameWindow
Prevented train head select dialog from being closed by clicking
outside the window

Everything necessary should be saved in order to avoid problems
when the onCreate is called.  Only thing I have yet to save is the
undo stacks, which I'll get to in the future.  I'll make hand
parcelable and let it handle all the hand information that needs
to be saved.

I made the train head 0 in both the constructers in handMT so
nothing is displayed in the train head area during new game
and set change transitions.